Basic macronutrient calculator
This allows you to provide an estimated TDEE (Total Daily Energy
Expenditure), a deficit percentage, or just the calories you want to eat
and then it will calculate the macronutrient goals/limits. This is great
to use in conjunction with MyFitnessPal or other calorie counting
applications.

The defaults here are for a Ketogenic diet. It's also possible to set a
hard limit on the number of carbs if so desired and then the calculator
will redistribute those to protein to ensure you don't lose too much
muscle mass.